Remove Float

+++ backend/db/migrations/20230122195438-create-spot.js
@@ -31,11 +31,11 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.FLOAT(9, 7),
+        type: Sequelize.FLOAT,
         defaultValue: 0.0
       },
       lng: {
-        type: Sequelize.FLOAT(10, 7),
+        type: Sequelize.FLOAT,
         defaultValue: 0.0
       },
       name: {
